Deduplicates ParDo{Single,Multi}EvaluatorFactory
This is in preparation for adding a third one for a future ParDo-like primitive
transform to be introduced inside SplittableParDo.
